+++ vscode/src/chat/chat-view/prompt.test.ts
@@ -22,7 +22,6 @@ describe('DefaultPrompter', () => {
             {
                 getExplicitContext: () => [],
                 getEnhancedContext: () => Promise.resolve([]),
-                getCommandContext: () => Promise.resolve([]),
             },
             true,
             100000
@@ -69,7 +68,6 @@ describe('DefaultPrompter', () => {
             {
                 getExplicitContext: () => [],
                 getEnhancedContext: () => Promise.resolve([]),
-                getCommandContext: () => Promise.resolve([]),
             },
             true,
             100000